refactor: headless export mode — no window for --export flag
- Added initVulkanHeadless() that skips window, surface, swapchain,
  and graphics pipeline — only creates compute essentials
- Fixed createLogicalDeviceHeadless() feature chain: replaced broken
  reinterpret_cast with proper StructureChain, fixing validation
  errors for synchronization2 and dynamicRendering
- Renamed createTexture() to createStarmapTexture() for clarity
- Minor cleanups: ~0 to UINT32_MAX, explicit static_casts,
  renamed local 'surface' variable to avoid shadowing

+++ include/utils.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <vulkan/vulkan.hpp>
 #include <GLFW/glfw3.h>
 #include <fstream>
@@ -57,5 +58,5 @@ inline fs::path get_executable_directory() {
         return fs::current_path(); // Fallback if readlink fails
     }
     // Returns the directory containing the binary (e.g., /path/to/project/build/)
-    return fs::path(std::string(result, count)).parent_path();
+    return fs::path(std::string(result, static_cast<size_t>(count))).parent_path();
 }
